refactor(utils): Move functions from `llm_benchmark` to `utils.py`
